+++ CHANGELOG.rst
@@ -21,11 +21,13 @@ Release date: ``2020-xx-xx``
 - `NXPY-174 <https://jira.nuxeo.com/browse/NXPY-174>`__: Improve ``test_repository.py`` reliability
 - `NXPY-176 <https://jira.nuxeo.com/browse/NXPY-176>`__: Add ``Nuxeo.can_use()`` to determine if a given operation is available
 - `NXPY-177 <https://jira.nuxeo.com/browse/NXPY-177>`__: Prevent ``AttributeError`` when fetching the server version and the response is bad (and return "unknown")
+- `NXPY-178 <https://jira.nuxeo.com/browse/NXPY-178>`__: Use a uniq ID for the S3 direct upload key
 
 Technical changes
 -----------------
 
 - ``Batch.complete()`` now handles additional parameters
+- Added ``Batch.key``
 - Added ``Nuxeo.can_use()``
 - Added ``Uploader.timeout()``
 - Added nuxeo/constants.py::\ ``LOG_LIMIT_SIZE``